+++ PyricetiFramework/Editor/Startup/ProjectData.cs
@@ -1,8 +1,0 @@
-﻿namespace PyricetiFramework.Editor
-{
-  public static class ProjectData
-  {
-    // ReSharper disable once StringLiteralTypo
-    public const string ProjectName = "superjeu";
-  }
-}