+++ PyricetiFramework/Runtime/Scripts/Data/ProjectData.cs
@@ -1,0 +1,7 @@
+﻿namespace PyricetiFramework
+{
+  public static class ProjectData
+  {
+    public static string ProjectName = "superjeu";
+  }
+}