+++ PyricetiFramework/Runtime/Scripts/Engine/EngineController.cs
@@ -1,3 +1,5 @@
+using UnityEngine;
+
 namespace PyricetiFramework
 {
   public abstract class EngineController : EngineObject
@@ -8,5 +10,14 @@ namespace PyricetiFramework
 
       ControllersProvider.registerController(this);
     }
+    
+    public static T getController<T>() where T : EngineController => ControllersProvider.getController<T>();
+  }
+
+  public abstract class EngineController<TSettings> : EngineController where TSettings : AbstractSettings
+  {
+    [SerializeField] protected TSettings settings = default;
+
+    public TSettings Settings => settings;
   }
 }